chore: extract eval data setup into shared module
Decouple the demo data creation logic (HedgeboxMatrix, core memory,
dashboard fixtures) from ee/hogai/eval/ci/conftest.py into a new shared
module ee/hogai/eval/data_setup.py. This allows the sandboxed agent
eval harness to reuse the same data setup without duplicating code.

- New file: ee/hogai/eval/data_setup.py with create_demo_org_team_user(),
  create_core_memory(), DashboardWithInsightsFixture, and
  create_dashboard_with_insights()
- Refactored ci/conftest.py to import from data_setup (no behavior change)
- Updated sandboxed/conftest.py with demo_org_team_user and core_memory
  fixtures using the shared functions
